Class file organization
Split separate classes out to new .py files for clarity of process.
