+++ libs/card.py
@@ -1,0 +1,8 @@
+class Card:
+    def __init__(self, name, value, suit):
+        self.name = name
+        self.value = value
+        self.suit = suit
+
+    def __str__(self):
+        return f"{self.name} of {self.suit}, worth {self.value} points"
